+++ src/components/RemoveBookButton.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
 import PropTypes from 'prop-types';
-import { removeBook } from '../redux/books/booksSlice';
+import { removeBook } from '../api/middleware';
 
 const RemoveBookButton = ({ bookId }) => {
   const dispatch = useDispatch();
