+++ src/redux/books/booksSlice.js
@@ -1,55 +1,5 @@
-import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
-import { v4 as uuidv4 } from 'uuid';
-import axios from 'axios';
-
-const apiKey = 'XbsVCoKrfytCMi4uh4fE';
-export const fetchBooks = createAsyncThunk(
-  'books/fetchBooks',
-  async () => {
-    try {
-      const response = await axios.get(`https://us-central1-bookstore-api-e63c8.cloudfunctions.net/bookstoreApi/apps/${apiKey}/books`);
-      return response.data;
-    } catch (error) {
-      throw new Error('Failed to fetch books.');
-    }
-  },
-);
-
-export const addBook = createAsyncThunk(
-  'books/addBook',
-  async (newBook) => {
-    try {
-      const bookToAdd = {
-        ...newBook,
-        item_id: uuidv4(),
-      };
-
-      const response = await axios.post(`https://us-central1-bookstore-api-e63c8.cloudfunctions.net/bookstoreApi/apps/${apiKey}/books`, bookToAdd);
-      console.log('Response from API:', response.data);
-      const addedBook = response.data;
-
-      return addedBook;
-    } catch (error) {
-      console.log('Error adding book:', error.response.status);
-      throw new Error('Failed to add book.');
-    }
-  },
-);
-
-export const removeBook = createAsyncThunk(
-  'books/removeBook',
-  async (bookId) => {
-    try {
-      console.log('removeBook thunk with bookId:', bookId);
-      await axios.delete(`https://us-central1-bookstore-api-e63c8.cloudfunctions.net/bookstoreApi/apps/${apiKey}/books/${bookId}`);
-      console.log('Book removed successfully:', bookId);
-      return { payload: bookId };
-    } catch (error) {
-      console.error('Error removing book:', error);
-      throw new Error('Failed to remove book.');
-    }
-  },
-);
+import { createSlice } from '@reduxjs/toolkit';
+import { fetchBooks, addBook, removeBook } from '../../api/middleware';
 
 const initialState = {
   books: [],
@@ -65,7 +15,7 @@ export const booksSlice = createSlice({
       const bookId = action.payload;
       const newBooks = { ...state.books };
       delete newBooks[bookId];
-      return { ...state, books: newBooks };
+      state.books = newBooks;
     },
   },
   extraReducers: (builder) => {
@@ -75,6 +25,7 @@ export const booksSlice = createSlice({
     builder.addCase(fetchBooks.fulfilled, (state, action) => {
       state.isLoading = false;
       state.books = action.payload;
+      console.log('Updated books state after fetchBooks:', state.books);
     });
     builder.addCase(fetchBooks.rejected, (state, action) => {
       state.isLoading = false;
@@ -86,6 +37,7 @@ export const booksSlice = createSlice({
     builder.addCase(addBook.fulfilled, (state, action) => {
       state.isLoading = false;
       const addedBook = action.payload;
+
       state.books[addedBook.item_id] = [addedBook];
     });
 
@@ -101,7 +53,8 @@ export const booksSlice = createSlice({
       const bookIdToDelete = action.payload;
       const newBooks = { ...state.books };
       delete newBooks[bookIdToDelete];
-      return { ...state, books: newBooks };
+      state.books = newBooks;
+      console.log('Updated books state after removeBook:', state.books);
     });
 
     builder.addCase(removeBook.rejected, (state, action) => {
